requirements.txt
PyGithub==2.6.1
azure-devops==7.1.0b4
python-gitlab==5.6.0